+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - [Feature: Optimized configuration update](https://github.com/Tencent/spring-cloud-tencent/pull/528)
 - [Feature：support pushGateway push metrics](https://github.com/Tencent/spring-cloud-tencent/pull/529)
 - [Feature: remove location metadata](https://github.com/Tencent/spring-cloud-tencent/pull/536)
+- [add feature-env plugin & add spring cloud gateway staining plugin](https://github.com/Tencent/spring-cloud-tencent/pull/533)